Moved GET request to a more appropriate spot
